added role change and approval functionality to api. added new data.sql with more inital data. refactored tests + controllers

+++ src/main/java/com/cathalob/medtracker/repository/BloodPressureReadingRepository.java
@@ -3,5 +3,5 @@ package com.cathalob.medtracker.repository;
 import com.cathalob.medtracker.model.tracking.BloodPressureReading;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BloodPressureReadingRepository extends JpaRepository<BloodPressureReading, Integer> {
+public interface BloodPressureReadingRepository extends JpaRepository<BloodPressureReading, Long> {
 }
